% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_01_apps.R
\name{add_app}
\alias{add_app}
\title{Polished API - Add an App}
\usage{
add_app(app_name, app_url = NULL, api_key = get_api_key())
}
\arguments{
\item{app_name}{the app name.}

\item{app_url}{an optional app url.  This url will be included in links sent out in invite
and email verification emails to redirect your users to your app.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  When successful, the \code{content} of the
\code{polished_api_res} is \code{list(message = "success")}.  In the case of an error, the
content is \code{list(error = "<error message>")}.
}
\description{
Polished API - Add an App
}
\seealso{
\code{\link[=get_apps]{get_apps()}} \code{\link[=update_app]{update_app()}} \code{\link[=delete_app]{delete_app()}}
}
