% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_06_sessions.R
\name{add_session}
\alias{add_session}
\title{Polished API - Add a session}
\usage{
add_session(add_uid, session_data, api_key = get_api_key())
}
\arguments{
\item{session_data}{list of data to include in the session.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}

\item{app_uid}{the app uid.}
}
\description{
Polished API - Add a session
}
