% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_05_user_roles.R
\name{add_user_role}
\alias{add_user_role}
\title{Polished API - Add a User Role}
\usage{
add_user_role(user_uid, role_uid, api_key = getOption("polished")$api_key)
}
\arguments{
\item{user_uid}{a user uid.}

\item{role_uid}{a role name.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Add a User Role
}
\seealso{
get_roles delete_role
}
