% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_button.R
\name{admin_button}
\alias{admin_button}
\title{admin_button}
\usage{
admin_button(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
The server logic for the button to navigate 'polished' admins from your 'shiny'
app to the 'polished' Admin Panel.
}
