% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_button.R
\name{admin_button_ui}
\alias{admin_button_ui}
\title{An html button to navigate the the "Admin Panel"}
\usage{
admin_button_ui(align = "right", vertical_align = "bottom")
}
\arguments{
\item{align}{The horizontal alignment of the button. Valid options are "right" (the default)
or "left".}

\item{vertical_align}{the vertical alignment of the button.  Valid options are "bottom" (the default)
or "top"}
}
\value{
admin button UI
}
\description{
The UI portion of the 'shiny' module for the button to navigate to the "Admin Panel".
This is the button that, when clicked, navigates a 'polished' admin from your 'shiny' app to the 'polished'
Admin Panel.  If your app is set up with the default 'polished' configuration, this button appears
in the bottom right of your 'shiny' app.
}
