% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_app.R
\name{admin_server}
\alias{admin_server}
\title{The server logic for the default Admin Panel dashboard}
\usage{
admin_server(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\value{
\code{invisible(NULL)}
}
\description{
The Shiny module server logic for the \code{polished} Admin Panel, accessible to Admin users.
}
