% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api.R
\name{api_list_to_df}
\alias{api_list_to_df}
\title{Convert a list returned from the Polished API into a data frame}
\usage{
api_list_to_df(api_list)
}
\arguments{
\item{api_list}{a list.  All elements in the list are vectors of the same length.}
}
\value{
a tibble
}
\description{
In order to avoid issues with converting R data frames into JSON objects and back
to R data frames, we instead convert R data frames to R lists before converting
them to JSON to be sent via the Polished API.  This function then converts those
lists back into R data frames (or more precisely tibbles).
}
