% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_filter.R
\name{auth_filter}
\alias{auth_filter}
\title{Auth filter for a Plumber API}
\usage{
auth_filter(req, res, method = "cookie")
}
\arguments{
\item{req}{the request}

\item{res}{the response}

\item{method}{The authentication method.  Valid options are "cookie" and/or "basic".  If
"cookie", the filter will authenticate the request using the cookie.  If
"basic" is set, the filter will authenticate the request using basic auth.  If both
"cookie" and "basic" are set, then the filter will first attempt to authenticate
using the cookie, and, if that fails, it will attempt to authenticate using basic
auth.}
}
\description{
Auth filter for a Plumber API
}
