% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_app.R
\name{bundle_app}
\alias{bundle_app}
\title{Create a tar archive}
\usage{
bundle_app(app_dir = ".")
}
\arguments{
\item{app_dir}{The path to the directory containing your Shiny app.  Defaults to the
working directory.}
}
\value{
the file path of the app bundle
}
\description{
This function is called by \code{deploy_app()} to compress Shiny apps before
deploying them to Polished Hosting.  You probably won't need to call this function
directly.
}
\examples{

\dontrun{
bundle_app(
  system.file("examples/polished_example_01", package = "polished")
)
}


}
