% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_timestamp}
\alias{convert_timestamp}
\title{convert_timestamp}
\usage{
convert_timestamp(timestamp, tz = "America/New_york")
}
\arguments{
\item{timestamp}{the JSON UTC date(s)}

\item{tz}{the timezone to convert the returned datetime to}
}
\value{
the R POSIX.ct datetime(s)
}
\description{
convert a Javascript JSON UTC date to a time in R.  We generally convert
Firestore timestamps to JSON UTC dates before we send them to R.
}
\examples{

json_times <- "2019-07-17T21:10:48.245Z"

convert_timestamp(json_times)

}
