% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_app_user.R
\name{create_app_user}
\alias{create_app_user}
\title{create_app_user}
\usage{
create_app_user(
  conn,
  app_uid,
  email,
  is_admin = FALSE,
  created_by = NULL,
  modified_by = NULL,
  schema = "polished"
)
}
\arguments{
\item{conn}{the database connection.}

\item{app_uid}{the name of the Shiny app.}

\item{email}{the email address of the first user.}

\item{is_admin}{boolean that defaults to FALSE.  Whether or not the user being created
is an admin.}

\item{created_by}{uid of the user creating this user.  If `NULL`, the default, then the
user uid of the user being created will be used.}

\item{modified_by}{uid of the user creating this user.  If `NULL`, the default, then the
value of `created_by` will be used.}

\item{schema}{the database schema}
}
\description{
Add the first user to the "polished" schema
}
