% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_first_user.R
\name{create_first_user}
\alias{create_first_user}
\title{create the first user of the Shiny app}
\usage{
create_first_user(firebase_functions_url, email, app_name,
  is_admin = TRUE, invite_status = "pending",
  time_created = lubridate::with_tz(Sys.time(), tzone = "UTC"))
}
\arguments{
\item{firebase_functions_url}{string - the Firebase functions url}

\item{email}{the first users email address}

\item{app_name}{string - the name of the app to add the user to}

\item{is_admin}{boolean - TRUE by default}

\item{invite_status}{string - "pending" by default}

\item{time_created}{POSIXct datetime - current time UTC by default}
}
\description{
Used to create the document that contains the first user of the Shiny app so that
the developer adding Polished to the Shiny app will be able to sign in without having
to manually create the user document in the Firestore web UI.
}
