% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uid.R
\name{create_uid}
\alias{create_uid}
\title{create a uid(s)}
\usage{
create_uid(n = 1)
}
\arguments{
\item{n}{length 1 integer > 0.  The number of uids to create.}
}
\value{
a character vector of uids
}
\description{
create a uid(s)
}
\examples{

# create a single uid
create_uid()

# create 10 uids
create_uid(10)

}
