% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module.R
\name{dashboard_module}
\alias{dashboard_module}
\title{dashboard_module}
\usage{
dashboard_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
dashboard_module
}
