% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_module.R
\name{dashboard_module_ui}
\alias{dashboard_module_ui}
\title{dashboard_module_ui}
\usage{
dashboard_module_ui(id)
}
\arguments{
\item{id}{the module id}
}
\value{
the UI for the "Dashboard" page of the "Admin Panel".
}
\description{
The 'shiny' module UI for the "Dashboard" page of the "Admin Panel".
}
