% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_app_user.R
\name{delete_app_user}
\alias{delete_app_user}
\title{delete app user}
\usage{
delete_app_user(conn, app_uid_, user_uid, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{app_uid_}{the app uid}

\item{user_uid}{the user uid}

\item{schema}{the database schema}
}
\description{
Delete an app user from the "app_users" table
}
