% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_04_roles.R
\name{delete_role}
\alias{delete_role}
\title{Polished API - Delete a Role}
\usage{
delete_role(role_uid, api_key = getOption("polished")$api_key)
}
\arguments{
\item{role_uid}{the role uid of the role to be deleted.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Delete a Role
}
\seealso{
\code{\link[=get_roles]{get_roles()}} \code{\link[=add_role]{add_role()}}
}
