% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_02_users.R
\name{delete_user}
\alias{delete_user}
\title{Polished API - Delete a User}
\usage{
delete_user(user_uid, api_key = get_api_key())
}
\arguments{
\item{user_uid}{the uid of the user to be deleted.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  When successful, the \code{content} of the
\code{polished_api_res} is \code{list(message = "success")}.  In the case of an error, the
content is \code{list(error = "<error message>")}.
}
\description{
Polished API - Delete a User
}
\seealso{
\code{\link[=get_users]{get_users()}} \code{\link[=add_user]{add_user()}} \code{\link[=update_user]{update_user()}}
}
