% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_05_user_roles.R
\name{delete_user_role}
\alias{delete_user_role}
\title{Polished API - Delete a User Role}
\usage{
delete_user_role(role_uid, user_uid, api_key = get_api_key())
}
\arguments{
\item{role_uid}{the role uid of the role to be deleted.}

\item{user_uid}{the user uid that the role should be removed from.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Delete a User Role
}
\seealso{
\code{\link[=get_user_roles]{get_user_roles()}} \code{\link[=add_user_role]{add_user_role()}}
}
