% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_input.R
\name{email_input}
\alias{email_input}
\title{shiny::textInput with type="email"}
\usage{
email_input(inputId, label, value = "", width = NULL, placeholder = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. '400px', or '100%'}

\item{placeholder}{A character string giving the user a hint as to what can be entered
into the control. Internet Explorer 8 and 9 do not support this option.}
}
\description{
set the "type" of the input tag to "email" rather than "text" to get the
correct email input browser behavior
}
