% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_init.R
\name{firebase_init}
\alias{firebase_init}
\title{firebase_init}
\usage{
firebase_init(firebase_config)
}
\arguments{
\item{firebase_config}{list of firebase configuration}
}
\value{
a character string of JavaScript code defining firebaseConfig
}
\description{
firebase_init
}
\examples{

my_config <- list(
  apiKey = "AIzaSyCTrjysW7-roaUOhGmJwvswh0KE4IYM3qk",
  authDomain = "polish-template.firebaseapp.com",
  databaseURL = "https://polish-template.firebaseio.com",
  projectId = "polish-template"
)

firebase_config(my_config)

}
