% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_rules.R
\name{generate_firestore_rules_text}
\alias{generate_firestore_rules_text}
\title{generate firestore rules text}
\usage{
generate_firestore_rules_text(app_names)
}
\arguments{
\item{app_names}{the names of all apps using firebase hosting}
}
\description{
This text will be used to write the firestore.rules file
}
