% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_firebase_hosting.R
\name{generate_hosting_text}
\alias{generate_hosting_text}
\title{generate firebase hosting text}
\usage{
generate_hosting_text(app_names)
}
\arguments{
\item{app_names}{the names of all apps using Firebase hosting}
}
\description{
This text is used as configuration in firebase.json
}
\examples{

generate_hosting_text("shiny_1")

generate_hosting_text(c("shiny_1", "shiny_2"))

}
