% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_active_users.R
\name{get_active_users}
\alias{get_active_users}
\title{get active users}
\usage{
get_active_users(conn, app_uid, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{app_uid}{the app id}

\item{schema}{the name of the schema}
}
\description{
get active users
}
