% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_users.R
\name{get_app_users}
\alias{get_app_users}
\title{get_app_users}
\usage{
get_app_users(conn, app_uid, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{app_uid}{the id of the app}

\item{schema}{the database schema}
}
\value{
a data frame of the app users from the polished schema.
}
\description{
get_app_users
}
