% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api.R
\name{get_apps}
\alias{get_apps}
\title{Polished API - Get App(s)}
\usage{
get_apps(
  app_uid = NULL,
  app_name = NULL,
  api_key = getOption("polished")$api_key
)
}
\arguments{
\item{app_uid}{an optional app uid.}

\item{app_name}{an optional app name.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Get App(s)
}
\details{
One of either the \code{app_uid} or \code{app_name} must not be \code{NULL}.  If both the
\code{app_uid} and \code{app_name} are provided, then the \code{app_uid} will be used, and the
\code{app_name} will be ignored.
}
\seealso{
add_apps update_app delete_app
}
