% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cookie}
\alias{get_cookie}
\title{get_cookie}
\usage{
get_cookie(cookie_string, name)
}
\arguments{
\item{cookie_string}{the cookie string}

\item{name}{the name of the cookie}
}
\description{
Get a cookie value by name from a cookie string
}
\examples{
cookies <- "cookie_name=cookie-value; cookie_name_2=cookie-value-2; cookie_name_3=cookie-with=sign"

polished:::get_cookie(cookies, "cookie_name")
polished:::get_cookie(cookies, "cookie_name_2")
polished:::get_cookie(cookies, "cookie_name_3")

}
