% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_sessions.R
\name{get_daily_sessions}
\alias{get_daily_sessions}
\title{get daily sessions}
\usage{
get_daily_sessions(conn, app_uid_, start_date, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{app_uid_}{the app id}

\item{start_date}{the start date for the query}

\item{schema}{the database schema}
}
\description{
get daily sessions
}
