% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_deps.R
\name{get_dependent_packages}
\alias{get_dependent_packages}
\title{get packages required to run R code}
\usage{
get_dependent_packages(directory = getwd())
}
\arguments{
\item{directory}{folder to search for R and Rmd files}
}
\value{
a vector of package names
}
\description{
Note: this function is copied from the \code{automagic} R package.  We are including it in
\code{polished} while we await the merging of this PR \url{https://github.com/cole-brokamp/automagic/pull/17}
and a new CRAN release of \code{automagic}.
}
\details{
parses all R and Rmd files in a directory and uses \code{automagic::parse_packages}
to find all R packages required for the code to run
}
