% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invite.R
\name{get_invite}
\alias{get_invite}
\title{get invite}
\usage{
get_invite(conn_, app_uid, user_uid, schema = "polished")
}
\arguments{
\item{conn_}{the database connection}

\item{app_uid}{the id of the app}

\item{user_uid}{the user uid}

\item{schema}{the name of the schema}
}
\description{
get invite
}
