% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_active_session_time.R
\name{get_last_active_session_time}
\alias{get_last_active_session_time}
\title{get last active session time}
\usage{
get_last_active_session_time(conn, app_uid_, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{app_uid_}{the uid of the app}

\item{schema}{the database schema}
}
\value{
a data frame with 2 columns:
- user_uid
- timestamp
The timestamps in this dataframe mark the most recent time that the user
has accessed the app.  Users without a row in this table have not yet accessed the
app
}
\description{
find the time that the last session became active for all users of a particular app
}
