% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_deps.R
\name{get_package_deps}
\alias{get_package_deps}
\title{Create a list of R Package Dependencies}
\usage{
get_package_deps(app_dir = ".", all_deps = FALSE)
}
\arguments{
\item{app_dir}{path to a directory containing R scripts or R Markdown files. Defaults
to current working directory.}

\item{all_deps}{boolean - whether or not to explicitly identify all "Imports" package dependencies
of the packages used by your app.}
}
\value{
a list of package dependencies with installation details
}
\description{
Given a \code{path} to a directory this function will scan all \code{.R} and
\code{.Rmd} files for any used R Packages along with their CRAN versions or GitHub references.
}
\details{
Currently, packages are detected via \code{\link[automagic:parse_packages]{automagic::parse_packages()}} which supports
the following calls within the code: \code{library()}, \code{require()}, and
prefixed \code{::} calls to functions.

Once an initial vector of package is created, each package is a checked that it can
be downloaded and installed from \code{CRAN} or \code{GitHub}.
}
\seealso{
\code{\link[automagic:parse_packages]{automagic::parse_packages()}}
}
\keyword{internal}
