% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api_04_roles.R
\name{get_roles}
\alias{get_roles}
\title{Polished API - Get Role(s)}
\usage{
get_roles(role_uid = NULL, api_key = getOption("polished")$api_key)
}
\arguments{
\item{role_uid}{an optional role uid.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  The "content" of the object is a
tibble of users(s) with the following columns:
\itemize{
\item uid
\item role_name
\item created_at
}
}
\description{
Polished API - Get Role(s)
}
\seealso{
add_role delete_role
}
