% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session.R
\name{get_session}
\alias{get_session}
\title{get session by hashed cookie}
\usage{
get_session(conn, hashed_cookie)
}
\arguments{
\item{conn}{the database connection}

\item{hashed_cookie}{the hashed cookie}
}
\value{
the signed in user session
}
\description{
get session by hashed cookie
}
