% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_06_sessions.R
\name{get_sessions}
\alias{get_sessions}
\title{Polished API - Get Sessions}
\usage{
get_sessions(app_uid, hashed_cookie, api_key = get_api_key())
}
\arguments{
\item{app_uid}{the app uid}

\item{hashed_cookie}{hashed cookie}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  The "content" of the object is a
tibble of users(s) with the following columns:
\itemize{
\item role_uid
\item role_name,
\item user_uid,
\item user_name,
\item created_at
}
}
\description{
This can currently only be used to get a single session
}
