% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_by_email.R
\name{get_user_by_email}
\alias{get_user_by_email}
\title{get user by email address}
\usage{
get_user_by_email(conn, account_uid, email, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{account_uid}{the account uid}

\item{email}{the user's email address}

\item{schema}{the database schema}
}
\value{
a list of user info if the user is found or `NULL`
}
\description{
get user by email address
}
