% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intlTelInputDependencies.R
\name{intlTelInputDependencies}
\alias{intlTelInputDependencies}
\title{Load dependencies for intl-tel-input for \code{phone_input}}
\usage{
intlTelInputDependencies(pkgVersion = "17.0.12")
}
\arguments{
\item{pkgVersion}{the version of the \code{intl-tel-input} package}
}
\description{
intlTelInputDependencies()
}
