% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_ui}
\alias{normalize_ui}
\title{normalize UI}
\usage{
normalize_ui(ui, request_)
}
\arguments{
\item{ui}{the Shiny ui}

\item{request_}{the request environment passed to the first argument of the UI
function}
}
\value{
the Shiny UI
}
\description{
the UI passed a shiny app can be a function HTML.  This function normalized the 2 different
formats so that they both use the character
}
