% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_input.R
\name{password_input}
\alias{password_input}
\title{A modification of 'shiny::passwordInput'}
\usage{
password_input(
  input_id,
  label = tagList(icon("unlock-alt"), "password"),
  value = "",
  style = "",
  placeholder = NULL
)
}
\arguments{
\item{input_id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this option.}

\item{width}{The width of the input, e.g. '400px', or '100%'.}
}
\description{
This modified version of shiny's passwordInput() does not actual send the password
to our shiny server.  It is just a regular password input that always keeps your
user's password on the client.  The password is used to sign the user in and then
converted to a JWT by Firebase, all on the client, before it is sent to your shiny
server.
}
