% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api.R
\name{polished_api_res}
\alias{polished_api_res}
\title{Send GET Request to the Polished API}
\usage{
polished_api_res(resp)
}
\arguments{
\item{resp}{a Polished API response}
}
\value{
an S3 object of class "polished_api_res".
}
\description{
Send GET Request to the Polished API
}
