% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pipeline.R
\name{print.polis_config}
\alias{print.polis_config}
\title{Print method for POLIS configuration}
\usage{
\method{print}{polis_config}(x, ...)
}
\arguments{
\item{x}{A \code{polis_config} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for POLIS configuration
}
\keyword{internal}
