% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{print.polis_project}
\alias{print.polis_project}
\title{Print a polis_project}
\usage{
\method{print}{polis_project}(x, ...)
}
\arguments{
\item{x}{A \link{polis_project}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a polis_project
}
