% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api_00.R
\name{print.polished_api_res}
\alias{print.polished_api_res}
\title{print polished_api_res}
\usage{
\method{print}{polished_api_res}(x, ...)
}
\arguments{
\item{x}{an S3 object of class \code{polished_api_res}.}

\item{...}{additional arguments.}
}
\description{
Generic print function for \code{polished_api_res} S3 class.
}
