% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_module.R
\name{profile_module}
\alias{profile_module}
\title{Profile Module Server}
\usage{
profile_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
The server logic to accompany the  \code{\link{profile_module_ui}}.
}
