% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_module.R
\name{profile_module_ui}
\alias{profile_module_ui}
\title{verify email page ui}
\usage{
profile_module_ui(id, firebase_config)
}
\arguments{
\item{firebase_config}{Firebase configuration}
}
\description{
verify email page ui
}
