% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_query_string}
\alias{remove_query_string}
\title{remove_query_string}
\usage{
remove_query_string(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{the Shiny session}
}
\description{
Remove the entire query string from the url.  This function should be called
inside the server function of your 'shiny' app.
}
