% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_password_module.R
\name{reset_password_module}
\alias{reset_password_module}
\title{server logic for Shiny module to reset password}
\usage{
reset_password_module(input, output, session, email)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}

\item{email}{A reactive value returning the email address to send the password
reset email to.}
}
\description{
This function sends s request to the polished.tech API to reset a user's
password.
}
