% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_password_module.R
\name{reset_password_module_ui}
\alias{reset_password_module_ui}
\title{the UI for a Shiny module to reset password}
\usage{
reset_password_module_ui(id)
}
\arguments{
\item{id}{the Shiny module id}
}
\description{
the UI for a Shiny module to reset password
}
