% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_render.R
\name{secure_render}
\alias{secure_render}
\title{Render and secure Rmarkdown document}
\usage{
secure_render(
  rmd_file_path,
  global_sessions_config_args = list(api_key = get_api_key()),
  sign_in_page_args = list(),
  sign_out_button = NULL
)
}
\arguments{
\item{rmd_file_path}{the path the to .Rmd file.}

\item{global_sessions_config_args}{arguments to be passed to \code{\link{global_sessions_config}}.}

\item{sign_in_page_args}{a named \code{list()} to customize the Sign In page
UI if values aren't included in YAML header. Valid names are \code{color}, \code{company_name},
\code{logo}, & \code{background_image}. (\strong{NOTE:} YAML header values override these values if both provided).}

\item{sign_out_button}{A Shiny \code{actionButton} or \code{actionLink} with \code{inputId = "sign_out"}.
If this argument is left as \code{NULL}, \code{secure_render} will attempt to add in an appropriate sign
out button/link depending on the output format of your .Rmd document.  Set this argument to \code{list()}
to not include a sign out button.}
}
\value{
a Shiny app object
}
\description{
\code{secure_render()} can be used to render and secure any Rmarkdown document.
Rendering is handled by \code{rmarkdown::render} and the then the rendered document
is secured with \code{polished} authentication.
}
\examples{

\dontrun{

secure_render(system.file("examples/rmds/flexdashboard.Rmd", package = "polished"))
secure_render(system.file("examples/rmds/flexdashboard_shiny.Rmd", package = "polished"))
secure_render(system.file("examples/rmds/html_document.Rmd", package = "polished"))
secure_render(system.file("examples/rmds/pdf_document.Rmd", package = "polished"))
}
}
