% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_static.R
\name{secure_static}
\alias{secure_static}
\title{Secure a static html page}
\usage{
secure_static(
  html_file_path,
  global_sessions_config_args,
  sign_out_button = shiny::actionButton("sign_out", "Sign Out", icon =
    shiny::icon("sign-out-alt"), style =
    "position: fixed; top: 15px; right: 15px; color: #FFFFFF; z-index: 9999; background-color: #0000FF; padding: 15px; border-radius: 20px; border-color: #0000FF")
)
}
\arguments{
\item{html_file_path}{the path the to html file.  See the details for more info.}

\item{global_sessions_config_args}{arguments to be passed to \code{\link{global_sessions_config}}.}

\item{sign_out_button}{action button with id "sign_out". Set to NULL to not include a sign out button.}
}
\value{
a Shiny app object
}
\description{
\code{secure_static()} can be used to secure any html
page using 'polished'.  It is often used to add 'polished' to ".Rmd" htmloutput
and flexdashboards.
}
\details{
To secure a static html page, place the html page in a folder named "www"
and call \code{secure_static()} from a file named "app.R".  The file structure should
look like:
\itemize{
\item app.R
\item www/
\itemize{
\item index.html
}
}

See an example here: \url{https://github.com/Tychobra/polished_example_apps/tree/master/05_flex_dashboard}
}
