% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api_00.R
\name{set_api_key}
\alias{set_api_key}
\title{set Polished API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{the Polished API key}
}
\value{
a list of the newly set polished R options
}
\description{
The API key is set as an R option at \code{getOption("polished")$api_key}.
}
\examples{

set_api_key(api_key = "<my Polished API key>")


}
