% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config_env.R
\name{set_config_env}
\alias{set_config_env}
\title{Automatically set the config environment}
\usage{
set_config_env(override = NULL)
}
\arguments{
\item{override}{Set the environment to "default" or "production" manually. \strong{CAUTION:}
Be sure you know the difference between "default" & "production" configuration environments.
Using the "production" environment will affect the database of the deployed application.}
}
\description{
Determines if the app is deployed to a server or running locally, and adjusts
the config environment to "production" or "default" respectively.  This function
is almost always called in the "global.R" file of a shiny app immediately before
the configuration in the "config.yml" is read in.
}
