% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_password_module.R
\name{set_password_module}
\alias{set_password_module}
\title{Shiny module server logic for verifying the user's email}
\usage{
set_password_module(input, output, session)
}
\arguments{
\item{input}{the shiny server input}

\item{output}{the shiny server output}

\item{session}{the shiny server session}
}
\description{
Shiny module server logic for verifying the user's email
}
