% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_password_module.R
\name{set_password_module_ui}
\alias{set_password_module_ui}
\title{Shiny module to set the user's password}
\usage{
set_password_module_ui(id)
}
\arguments{
\item{id}{the shiny module id}
}
\description{
The user is redirected to this page after they click their email invite.  They
set their new password on this page, and sign in to the Shiny app.
}
