% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_email.R
\name{sign_in_email}
\alias{sign_in_email}
\title{sign in vial email password}
\usage{
sign_in_email(email, password, hashed_cookie = create_cookie())
}
\arguments{
\item{email}{the user's email address}

\item{password}{the user's password}

\item{cookie}{the hashed_cookie}
}
\description{
sign in vial email password
}
