% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_components.R
\name{sign_in_js}
\alias{sign_in_js}
\title{Sign in and register pages JavaScript dependencies}
\usage{
sign_in_js(ns = function(x) x)
}
\arguments{
\item{ns}{the ns function from the Shiny module that this function is called
within.}
}
\description{
This function should be called at the bottom of your custom sign in and registration
pages UI.  It loads in all the JavaScript dependencies to handle polished sign
in and registration.  See the vignette for details.
}
