% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_ui.R
\name{sign_in_module}
\alias{sign_in_module}
\title{sign_in}
\usage{
sign_in_module(input, output, session, conn)
}
\arguments{
\item{input}{the Shiny input}

\item{output}{the Shiny output}

\item{session}{the Shiny session}

\item{conn}{a database connection}
}
\description{
sign_in
}
