% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_module_2.R
\name{sign_in_module_2}
\alias{sign_in_module_2}
\title{Server logic for the Sign In & Register pages}
\usage{
sign_in_module_2(input, output, session)
}
\arguments{
\item{input}{the Shiny \code{input}}

\item{output}{the Shiny \code{output}}

\item{session}{the Shiny \code{session}}
}
\description{
This server logic accompanies \code{\link{sign_in_module_2_ui}}.
}
