% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_module_2.R
\name{sign_in_module_ui_2}
\alias{sign_in_module_ui_2}
\title{UI for the sign in and register pages}
\usage{
sign_in_module_ui_2(id)
}
\arguments{
\item{id}{the Shiny module id}
}
\description{
Alternate sign in UI that works regardless of whether or not invites
are required. \code{\link{sign_in_module_2}} must be provided as the 
argument custom_sign_in_server in \code{\link{secure_server}} for proper
functionality.
}
