% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_no_invite_module.R
\name{sign_in_no_invite_module}
\alias{sign_in_no_invite_module}
\title{sign_in_no_invite_module}
\usage{
sign_in_no_invite_module(input, output, session)
}
\arguments{
\item{input}{the Shiny input}

\item{output}{the Shiny output}

\item{session}{the Shiny session}
}
\description{
sign_in_no_invite_module
}
