% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_ui_default.R
\name{sign_in_ui_default}
\alias{sign_in_ui_default}
\title{sign_in_ui_default}
\usage{
sign_in_ui_default(firebase_config)
}
\arguments{
\item{firebase_config}{Firebase configuration}
}
\value{
the UI for the sign in page
}
\description{
sign_in_ui_default
}
