% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_01_apps.R
\name{update_app}
\alias{update_app}
\title{Polished API - Update an App}
\usage{
update_app(app_uid, app_name = NULL, app_url = NULL, api_key = get_api_key())
}
\arguments{
\item{app_uid}{the app uid of the app to update.}

\item{app_name}{an optional app name to replace the existing app name.}

\item{app_url}{an optional app url to replace the existing app url.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  When successful, the \code{content} of the
\code{polished_api_res} is \code{list(message = "success")}.  In the case of an error, the
content is \code{list(error = "<error message>")}.
}
\description{
Polished API - Update an App
}
\seealso{
\code{\link[=get_apps]{get_apps()}} \code{\link[=add_app]{add_app()}} \code{\link[=delete_app]{delete_app()}}
}
