% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_06_sessions.R
\name{update_session}
\alias{update_session}
\title{Polished API - Update a session}
\usage{
update_session(session_uid, session_data, api_key = get_api_key())
}
\arguments{
\item{session_uid}{the uid of the session to be updated.}

\item{session_data}{list of data to include in the updated session.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Update a session
}
\seealso{
\code{\link[=get_sessions]{get_sessions()}} \code{\link[=add_session]{add_session()}}
}
