% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\docType{data}
\name{User}
\alias{User}
\title{User}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
User
}
\description{
R6 class to track the polish user.  An instance of this class named `polish__user` should
be created in "global.R" of the Shiny app.
}
\examples{

user <- User$new(
  "https://us-central1-gatsby-firebase-5adaa.cloudfunctions.net/signInWithToken",
  "eyJhbGciOiJSUzI1NiIsImtpZCI6IjY2NDNkZDM5ZDM4ZGI4NWU1NjAxN2E2OGE3NWMyZjM4YmUxMGM1MzkiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL3NlY3VyZXRva2VuLmdvb2dsZS5jb20vcG9saXNoLXRlbXBsYXRlIiwiYXVkIjoicG9saXNoLXRlbXBsYXRlIiwiYXV0aF90aW1lIjoxNTU4NDkwNzA2LCJ1c2VyX2lkIjoiNUlseEpHNENnYU03M0lka3lDWDBEeGRxZVNoMSIsInN1YiI6IjVJbHhKRzRDZ2FNNzNJZGt5Q1gwRHhkcWVTaDEiLCJpYXQiOjE1NTg5NjMxMjYsImV4cCI6MTU1ODk2NjcyNiwiZW1haWwiOiJhbmR5Lm1lcmxpbm9AdHljaG9icmEuY29tIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsImZpcmViYXNlIjp7ImlkZW50aXRpZXMiOnsiZW1haWwiOlsiYW5keS5tZXJsaW5vQHR5Y2hvYnJhLmNvbSJdfSwic2lnbl9pbl9wcm92aWRlciI6InBhc3N3b3JkIn19.eM1fvVBgIybL1X4Spq3Kwzt3EQNSqi9I8njXxf_1OEvgY7zRrYiyxNjCaEhDbdpeTN7rX84YldN8PtfBgsO9BXx0XAZsAqpuWkj5U2pKDHddzr2bk"
)

user$get_is_authed()

user$sign_in_with_token("invalid_token")

# The below line of code will only work while the below token is still valid.
#user$sign_in_with_token("eyJhbGciOiJSUzI1NiIsImtpZCI6IjY2NDNkZDM5ZDM4ZGI4NWU1NjAxN2E2OGE3NWMyZjM4YmUxMGM1MzkiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL3NlY3VyZXRva2VuLmdvb2dsZS5jb20vcG9saXNoLXRlbXBsYXRlIiwiYXVkIjoicG9saXNoLXRlbXBsYXRlIiwiYXV0aF90aW1lIjoxNTU4NDkwNzA2LCJ1c2VyX2lkIjoiNUlseEpHNENnYU03M0lka3lDWDBEeGRxZVNoMSIsInN1YiI6IjVJbHhKRzRDZ2FNNzNJZGt5Q1gwRHhkcWVTaDEiLCJpYXQiOjE1NTg5NjMxMjYsImV4cCI6MTU1ODk2NjcyNiwiZW1haWwiOiJhbmR5Lm1lcmxpbm9AdHljaG9icmEuY29tIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsImZpcmViYXNlIjp7ImlkZW50aXRpZXMiOnsiZW1haWwiOlsiYW5keS5tZXJsaW5vQHR5Y2hvYnJhLmNvbSJdfSwic2lnbl9pbl9wcm92aWRlciI6InBhc3N3b3JkIn19.eM1fvVBgIybL1X4Spq3Kwzt3EQNSqi9I8njXxf_1OEvgY7zRrYiyxNjCaEhDbdpeTN7rX84YldN8PtfBgsO9BXx0XAZsAqpuWkj5U2pKDHddzr2bk-797o4Ppgknvp4rZcQCoW2E5B15YJ1ksHNZHLoejp7bKC83p4XThoCIjbaHzDik8-rlN8WUUrSPD0asG7CRhjNhJOIBQlphbLFhYqwTRid-XsgALG8SSjK-oT0ZzXJ0b3q7Wm2nUgbAkAUyErYlTO-WzScFUdjWLvvIhqjzeTD2Ll58zfJGSfqOuRMLUXJ1NmeARl-2cMw0gpN_HA-VSu_BgH1RR0VZFqehVA")

}
\keyword{datasets}
