% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_access_module.R
\name{user_access_module}
\alias{user_access_module}
\title{admin user access module}
\usage{
user_access_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\value{
\code{invisible(NULL)}
}
\description{
Server function for the default Shiny module to control user access in the \code{polished}
Admin Panel.
}
