% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_access_module.R
\name{user_access_module_ui}
\alias{user_access_module_ui}
\title{admin user access_ui}
\usage{
user_access_module_ui(id)
}
\arguments{
\item{id}{the module id}
}
\value{
the UI to create the user access module.
}
\description{
Shiny module UI for the default user access tab in the \code{polished} Admin Panel.
}
