% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_edit_module.R
\name{user_edit_module}
\alias{user_edit_module}
\title{user_edit_module}
\usage{
user_edit_module(
  input,
  output,
  session,
  modal_title,
  user_to_edit,
  open_modal_trigger,
  existing_roles,
  existing_users
)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny sever function output}

\item{session}{Shiny server function session}

\item{modal_title}{the title for the modal}

\item{user_to_edit}{reactive - a one row data frame of the user to edit from the "app_users" table.}

\item{open_modal_trigger}{reactive - a trigger to open the modal}

\item{existing_roles}{reactive data frame of all roles for this app}

\item{existing_users}{reactive data frame of all users of this app.  This is used to check that the user
does not add a user that already exists.}
}
\description{
user_edit_module
}
