% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_app.R
\docType{data}
\name{valid_gcp_regions}
\alias{valid_gcp_regions}
\title{Valid Regions for Polished Hosting}
\format{
An object of class \code{character} of length 35.
}
\usage{
valid_gcp_regions
}
\description{
Set the \code{region} argument of \code{deploy_app()} to one of these regions.
}
\keyword{datasets}
