% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_email_module.R
\name{verify_email_module}
\alias{verify_email_module}
\title{verify email page ui}
\usage{
verify_email_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
verify email page ui
}
