% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_firebase_hosting.R
\name{write_firebase_hosting}
\alias{write_firebase_hosting}
\title{write firebase hosting configuration to firebase.json}
\usage{
write_firebase_hosting(app_names, file_path = "firebase.json")
}
\arguments{
\item{app_names}{The name(s) of the apps used}

\item{file_path}{The path of the firebase hosting configuration file}
}
\description{
write firebase hosting configuration to firebase.json
}
\examples{

write_firebase_hosting("shiny_app_1")

write_firebase_hosting(c("shiny_app_1", "shiny_app_2"))

}
