% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_hosting_html.R
\name{write_firebase_hosting_html}
\alias{write_firebase_hosting_html}
\title{write firebase hosting html}
\usage{
write_firebase_hosting_html(app_names)
}
\arguments{
\item{app_names}{the names of the apps}
}
\description{
write firebase hosting html for one or more apps
}
\examples{

write_firebase_hosting_html("shiny_app_1")

write_firebase_hosting_html(c("shiny_app_1", "shiny_app_2"))

}
