% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_hosting_html.R
\name{write_firebase_hosting_html_single_app}
\alias{write_firebase_hosting_html_single_app}
\title{write hosting html to a file}
\usage{
write_firebase_hosting_html_single_app(app_name)
}
\arguments{
\item{app_name}{the name of the app}
}
\description{
write hosting html to a file
}
\examples{

write_firebase_hosting_html_single_app("shiny_app_1")

}
