% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_rules.R
\name{write_firestore_rules}
\alias{write_firestore_rules}
\title{write the firestore.rules file}
\usage{
write_firestore_rules(app_names, dir = ".",
  file_name = "firestore.rules")
}
\arguments{
\item{app_names}{The name(s) of the apps used}

\item{dir}{The directory for firestore.rules}

\item{file_name}{The name of the new file}
}
\description{
write the firestore.rules file
}
\examples{

write_firestore_rules("project1")

write_firestore_rules(c("project1", "project2"), "R", "file.rules")

}
