% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_app.R
\name{add_app}
\alias{add_app}
\title{add_app}
\usage{
add_app(conn, app_uid, app_name, created_by, modified_by, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{app_uid}{the app uid}

\item{app_name}{the app name}

\item{created_by}{user_uid for user who created the app}

\item{modified_by}{user_uid for the user who modified the app}

\item{schema}{the database schema}
}
\value{
number of rows affected or an error
}
\description{
add_app
}
