% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_app.R
\name{add_app}
\alias{add_app}
\title{add_app}
\usage{
add_app(conn, account_uid, app_uid, app_name, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{account_uid}{the account uid}

\item{app_uid}{the app uid}

\item{app_name}{the app name}

\item{schema}{the database schema}
}
\value{
number of rows affected or an error
}
\description{
add_app
}
