% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_01_apps.R
\name{add_app}
\alias{add_app}
\title{Polished API - Add an App}
\usage{
add_app(app_name, app_url = NULL, api_key = getOption("polished")$api_key)
}
\arguments{
\item{app_name}{the app name.}

\item{app_url}{an optional app url.  This url will be included in links sent out in invite
and email verification emails to redirect your users to your app.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Add an App
}
\seealso{
get_apps update_app delete_app
}
