% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api_03_app_users.R
\name{add_app_user}
\alias{add_app_user}
\title{Polished API - Add a User to an App}
\usage{
add_app_user(
  app_uid,
  user_uid = NULL,
  email = NULL,
  is_admin = FALSE,
  send_invite_email = FALSE,
  api_key = getOption("polished")$api_key
)
}
\arguments{
\item{app_uid}{an optional app uid.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}

\item{app_name}{an optional app name.}
}
\description{
Polished API - Add a User to an App
}
\details{
supply either the app uid or app name to get data about a specific app.
}
\seealso{
get_apps update_app delete_app
}
