% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_filter.R
\name{add_auth_to_spec}
\alias{add_auth_to_spec}
\title{add_auth_to_spec}
\usage{
add_auth_to_spec(method = c("basic", "cookie"))
}
\arguments{
\item{method}{the authentication method(s)}
}
\description{
Add authentication to the openapi plumber spec so that you can use the swagger
documentation with the \code{auth_filter()}.
}
\details{
This minimal API example \url{https://github.com/Tychobra/polished_example_apps/blob/master/11_plumber/api/00_start.R}
shows how you can add this function to your API.
}
