% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_04_roles.R
\name{add_role}
\alias{add_role}
\title{Polished API - Add a Role}
\usage{
add_role(role_name, api_key = getOption("polished")$api_key)
}
\arguments{
\item{role_name}{a role name..}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Add a Role
}
\seealso{
get_roles delete_role
}
