% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_session.R
\name{add_session}
\alias{add_session}
\title{add session to the "sessions" table}
\usage{
add_session(conn_, session, app_name)
}
\arguments{
\item{conn_}{the database connection}

\item{session}{list containing session data with the following elements:
- session_uid
- user_uid
- email
- email_verified
- hashed_cookie}

\item{app_name}{the name of the apps}
}
\description{
add session to the "sessions" table
}
