% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user.R
\name{add_user}
\alias{add_user}
\title{add_user}
\usage{
add_user(conn_, email, created_by)
}
\arguments{
\item{conn_}{the database connection}

\item{email}{the email address of the user to be added}

\item{created_by}{the uid of the user that created this new user}
}
\description{
add_user
}
