% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_02_users.R
\name{add_user}
\alias{add_user}
\title{Polished API - Add a User}
\usage{
add_user(email, api_key = getOption("polished")$api_key)
}
\arguments{
\item{email}{an email address.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Add a User
}
\seealso{
\code{\link[=get_users]{get_users()}} \code{\link[=delete_user]{delete_user()}}
}
