% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_button.R
\name{admin_button_ui}
\alias{admin_button_ui}
\title{admin_button_ui}
\usage{
admin_button_ui(id)
}
\arguments{
\item{id}{the Shiny module id.}
}
\value{
admin button UI
}
\description{
The default UI for the admin button 'shiny' module.  This is the button that,
when clicked, navigates a 'polished' admin from your 'shiny' app to the 'polished'
Admin Panel.
}
