% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_module.R
\name{admin_module}
\alias{admin_module}
\title{The server logic for the default Admin Panel dashboard}
\usage{
admin_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
The Shiny module server logic for the \code{polished} Admin Panel, accessible to Admin users.
}
