% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_module.R
\name{admin_module}
\alias{admin_module}
\title{admin_module}
\usage{
admin_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server input}

\item{output}{the Shiny server output}

\item{session}{the Shiny server session}
}
\description{
admin_module
}
