% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_module.R
\name{admin_module_ui}
\alias{admin_module_ui}
\title{admin_ui}
\usage{
admin_module_ui(id, firebase_config)
}
\arguments{
\item{id}{the Shiny module id}

\item{firebase_config}{list of Firebase config}
}
\description{
admin_ui
}
