% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_filter.R
\name{auth_filter}
\alias{auth_filter}
\title{Auth filter for a Plumber API}
\usage{
auth_filter(method = c("basic", "cookie"), api_key = get_api_key())
}
\arguments{
\item{method}{The authentication method.  Valid options are "basic" and/or "cookie".  If
"basic" is set, the filter will authenticate the request using basic auth.  If
"cookie", the filter will authenticate the request using the cookie.  If both
"cookie" and "basic" are set, then the filter will first attempt to authenticate
using the cookie, and, if that fails, it will attempt to authenticate using basic
auth.  If you use cookie based auth, and you want to send requests directly from the browser,
then be sure to set your Plumber API to allow for cookies.  See
\url{https://polished.tech/blog/polished-plumber} for details.}

\item{api_key}{Your polished API key}
}
\description{
Auth filter for a Plumber API
}
