% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_filter.R
\name{auth_filter}
\alias{auth_filter}
\title{Auth filter for a Plumber API}
\usage{
auth_filter(req, res)
}
\arguments{
\item{req}{the request}

\item{res}{the response}
}
\description{
Auth filter for a Plumber API
}
