% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_first_user}
\alias{create_first_user}
\title{create_first_user}
\usage{
create_first_user(conn, app_name, email)
}
\arguments{
\item{conn}{the database connection.}

\item{app_name}{the name of the Shiny app.}

\item{email}{the email address of the first user.}
}
\description{
Add the first user to the "polished" schema
}
