% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_schema}
\alias{create_schema}
\title{create_schema}
\usage{
create_schema(conn)
}
\arguments{
\item{conn}{the `RPostgres` database connection.  Only `RPostgres` is supported.}
}
\description{
create all the tables in the "polished" schema.  The "polished" schema contains the following tables.
 - users
 - apps
 - app_users
 - roles
 - user_roles
 - sessions
}
