% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_admin_ui_options.R
\name{default_admin_ui_options}
\alias{default_admin_ui_options}
\title{Default Options for the Admin UI}
\usage{
default_admin_ui_options()
}
\value{
the default list of HTML for branding elements in the Admin Panel UI. The valid list element names are:
\itemize{
  \item \code{title} - Title/Logo element in top left corner of Admin Panel dashboard & browser tab title
  \item \code{sidebar_branding} - Branding (e.g. Logo) on left sidebar of Admin Panel dashboard
  \item \code{browser_tab_icon} - Icon to display in browser tab
}
}
\description{
This function specifies the default logos that are displayed in the "Admin Panel".
}
