% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_admin_ui_options.R
\name{default_admin_ui_options}
\alias{default_admin_ui_options}
\title{Default Options for the Admin UI}
\usage{
default_admin_ui_options()
}
\value{
a list of html for branding the Admin UI
}
\description{
This function specifies the default logos that are displayed in the "Admin Panel".
}
