% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_01_apps.R
\name{delete_app}
\alias{delete_app}
\title{Polished API - Delete an App}
\usage{
delete_app(app_uid, api_key = getOption("polished")$api_key)
}
\arguments{
\item{app_uid}{the app uid.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Delete an App
}
\seealso{
\code{\link[=get_apps]{get_apps()}} \code{\link[=add_app]{add_app()}} \code{\link[=update_app]{update_app()}}
}
