% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_03_app_users.R
\name{delete_app_user}
\alias{delete_app_user}
\title{Polished API - Delete an App User}
\usage{
delete_app_user(app_uid, user_uid, api_key = get_api_key())
}
\arguments{
\item{app_uid}{an app uid.}

\item{user_uid}{a user uid.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  When successful, the \code{content} of the
\code{polished_api_res} is \code{list(message = "success")}.  In the case of an error, the
content is \code{list(error = "<error message>")}.
}
\description{
Polished API - Delete an App User
}
\seealso{
\code{\link[=get_apps]{get_apps()}} \code{\link[=add_app]{add_app()}} \code{\link[=update_app_user]{update_app_user()}}
}
