% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api.R
\name{delete_user}
\alias{delete_user}
\title{Polished API - Delete a User}
\usage{
delete_user(user_uid, api_key = getOption("polished")$api_key)
}
\arguments{
\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}

\item{app_uid}{the user uid of the user to be deleted.}
}
\description{
Polished API - Delete a User
}
\seealso{
get_apps add_app update_app
}
