% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_app.R
\name{deploy_app}
\alias{deploy_app}
\title{Deploy a Shiny app to Polished Hosting}
\usage{
deploy_app(
  app_name,
  app_dir = ".",
  api_key = get_api_key(),
  launch_browser = TRUE,
  region = "us-east1",
  ram_gb = 2,
  r_ver = NULL,
  tlmgr = character(0),
  golem_package_name = NULL,
  cache = TRUE,
  gh_pat = NULL
)
}
\arguments{
\item{app_name}{Your Shiny app's name.}

\item{app_dir}{The path to the directory containing your Shiny app.}

\item{api_key}{Your \code{polished} API key. Defaults to \code{Sys.getenv("POLISHED_API_KEY")} if set.}

\item{launch_browser}{Boolean (default: \code{TRUE}) - Whether or not to open
your newly deployed app in your default web browser after successful deployment.}

\item{region}{the region to deploy the app to on Google Cloud Platform. See
\url{https://cloud.google.com/run/docs/locations} for all available regions
on Google Cloud Platform. Currently, database connections are only supported for
\code{us-east1}. See \url{https://polished.tech/docs/06-database-connections} for details.}

\item{ram_gb}{the amount of memory (in \code{GiB}) to allocate to your Shiny app's server.
Valid values are \code{2}, \code{4}, \code{8}, \code{16}, or \code{32}.}

\item{r_ver}{Character string of desired \code{R} version.  If kept as \code{NULL} (the default),
\code{deploy_app()} will detect the R version you are currently running.  The R version must be a version
supported by an \code{r-ver} Docker image.  You can see all the \code{r-ver} Docker image versions
of R here \url{https://github.com/rocker-org/rocker-versioned2/tree/master/dockerfiles} and here
\url{https://github.com/rocker-org/rocker-versioned/tree/master/r-ver}.}

\item{tlmgr}{a character vector of \verb{TeX Live} packages to install.  This is only used if your Shiny
app generates \code{PDF} documents.  Defaults to \code{character(0)} for no \verb{TeX Live} installation. Provide a
character vector of your TeX Live package dependencies to have all your TeX Live packages installed at build time.}

\item{golem_package_name}{if your Shiny app was created as a package with the
\code{golem} package, provide the name of the Shiny app package as a character string.
Defaults to \code{NULL}.  Keep as \code{NULL} for non \code{golem} Shiny apps.}

\item{cache}{Boolean (default: \code{TRUE}) - whether or not to cache the Docker image.}

\item{gh_pat}{optional GitHub PAT for installing packages from private GitHub repos.}
}
\description{
Deploy a Shiny app to Polished Hosting
}
\examples{

\dontrun{
deploy_app(
  app_name = "polished_example_01",
  app_dir = system.file("examples/polished_example_01", package = "polished"),
  api_key = "<your polished.tech API key>"
)
}


}
