% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_dependencies.R
\name{firebase_dependencies}
\alias{firebase_dependencies}
\title{Load the Firebase JavaScript dependencies into the UI}
\usage{
firebase_dependencies(services = c("auth"), firebase_version = "7.15.5")
}
\arguments{
\item{services}{character vector of Firebase services to load into the UI.  Valid strings are
"auth", "firestore", "functions", "messaging", and "storage"}

\item{firebase_version}{character string of the Firebase version.  Defaults to 7.15.5.}
}
\value{
the html <script> tags for the Firebase JavaScript dependencies
}
\description{
Under the hood, 'polished' uses Firebase JavaScript dependencies to handle user
user authentication.  This function loads the required Firebase JavaScript dependencies
in the the UI of your 'shiny' app.
}
\examples{

firebase_dependencies()

}
