% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_init.R
\name{firebase_init}
\alias{firebase_init}
\title{Initialize Firebase}
\usage{
firebase_init(firebase_config)
}
\arguments{
\item{firebase_config}{list of firebase configuration}
}
\value{
a character string of JavaScript code to initialize Firebase
}
\description{
Executes a couple lines of JavaScript to initialize Firebase.  This function should be
called in your 'shiny' UI immediately after \code{\link{firebase_dependencies}}.
}
\examples{


\dontrun{
my_config <- list(
  apiKey = "your Firebase API key",
  authDomain = "your Firebase auth domain",
  projectId = "your Firebase Project ID"
)

firebase_init(my_config)
}
}
