% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_init.R
\name{firebase_init}
\alias{firebase_init}
\title{firebase_init}
\usage{
firebase_init(firebase_config)
}
\arguments{
\item{firebase_config}{list of firebase configuration}
}
\value{
a character string of JavaScript code defining firebaseConfig
}
\description{
firebase_init
}
\examples{

\dontrun{
my_config <- list(
  apiKey = "your Firebase API key",
  authDomain = "your Firebase auth domain",
  databaseURL = "your Firebase database URL",
  projectId = "your Firebase Project ID"
)

firebase_init(my_config)
}
}
