% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_hosting_html.R
\name{generate_firebase_hosting_html}
\alias{generate_firebase_hosting_html}
\title{html for iframe for hosting the Shiny app}
\usage{
generate_firebase_hosting_html(app_name, title = NULL)
}
\arguments{
\item{app_name}{the name of a single Shiny app}

\item{file_name}{default to "hosting/index.html".  The file name of the generated html.}
}
\description{
html for iframe for hosting the Shiny app
}
\examples{

write_firebase_hosting_html("shiny_app_1")

}
