% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_03_app_users.R
\name{get_app_users}
\alias{get_app_users}
\title{Polished API - Get App(s) User(s)}
\usage{
get_app_users(
  app_uid = NULL,
  user_uid = NULL,
  email = NULL,
  api_key = getOption("polished")$api_key
)
}
\arguments{
\item{app_uid}{an app uid.}

\item{user_uid}{an optional user uid for the user to be invited to the app.}

\item{email}{an optional email address for the user to be invited to the app.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{polished_api_res}.  The "content" of the object is a
tibble of app(s) with the following columns:
\itemize{
\item uid
\item app_name
\item app_url
\item created_at
\item modified_at
}
}
\description{
Polished API - Get App(s) User(s)
}
\seealso{
add_app update_app delete_app
}
