% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cookie}
\alias{get_cookie}
\title{get_cookie}
\usage{
get_cookie(cookie_string, name)
}
\arguments{
\item{cookie_string}{the cookie string}

\item{name}{the name of the cookie}
}
\description{
Get a cookie value by name from a cookie string
}
\examples{
cookies <- "_ga=GA1.1.768093326.1554508951; PGADMIN_KEY=a20922d4-e598-42e8-b62c-a9c9f833d249; PGADMIN_LANGUAGE=en"

get_cookie(cookies, "_ga")

}
