% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_active_session_time.R
\name{get_last_active_session_time}
\alias{get_last_active_session_time}
\title{get last active session time}
\usage{
get_last_active_session_time(conn, app_name_)
}
\arguments{
\item{conn}{the database connection}
}
\description{
get last active session time
}
