% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session.R
\name{get_session}
\alias{get_session}
\title{get session by hashed cookie}
\usage{
get_session(conn, account_uid, hashed_cookie, app_uid, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{account_uid}{the account uid}

\item{hashed_cookie}{the hashed cookie}

\item{app_uid}{the id of the app}

\item{schema}{the database schema}
}
\value{
the signed in user session
}
\description{
get session by hashed cookie
}
