% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_by_email.R
\name{get_user_by_email}
\alias{get_user_by_email}
\title{get user by email address}
\usage{
get_user_by_email(conn_, email)
}
\arguments{
\item{conn_}{the database connection}

\item{email}{the user's email address}
}
\value{
a list of user info if the user is found or `NULL`
}
\description{
get user by email address
}
