% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_sessions_config.R
\name{global_sessions_config}
\alias{global_sessions_config}
\title{configuration for global sessions}
\usage{
global_sessions_config(app_name, firebase_functions_url)
}
\arguments{
\item{app_name}{the name of the app.}

\item{firebase_functions_url}{the url of the firebase functions.}
}
\description{
configuration for global sessions
}
