% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_sessions_config.R
\name{global_sessions_config}
\alias{global_sessions_config}
\title{configuration for global sessions}
\usage{
global_sessions_config(
  app_name,
  firebase_project_id,
  conn,
  authorization_level = "app"
)
}
\arguments{
\item{app_name}{the name of the app.}

\item{firebase_project_id}{the Firebase project ID.}

\item{conn}{the database connection}

\item{authorization_level}{either "app" or "all".  Use "app" to individually authorize users to this
app.  Use "all" to give all user that have access to any of your apps access to this app.  "all" is
used by our "apps_dashboard_*".}
}
\description{
configuration for global sessions
}
