% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_sessions_config.R
\name{global_sessions_config}
\alias{global_sessions_config}
\title{configuration for global sessions}
\usage{
global_sessions_config(
  conn = NULL,
  app_name = NULL,
  firebase_config = NULL,
  admin_mode = FALSE,
  is_invite_required = TRUE,
  api_key = NULL
)
}
\arguments{
\item{conn}{the database connection}

\item{app_name}{the name of the app.}

\item{firebase_config}{a list containing your Firebase project configuration.  This list should have the
following named elements:
\itemize{
  \item{apiKey}
  \item{authDomain}
  \item{projectId}
}}

\item{admin_mode}{FALSE by default.  Set to TRUE to enter the polished Admin Panel without needing
to register and sign in.  This is useful during development for inviting the first users to your app.
Make sure to set `admin_mode` to FALSE before deploying your app.}

\item{is_invite_required}{TRUE by default.  Whether or not to require the user to have an
invite before registering/signing in}

\item{api_key}{the API key from Polished hosted}
}
\description{
configuration for global sessions
}
