% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_sessions_config.R
\name{global_sessions_config}
\alias{global_sessions_config}
\title{configuration for global sessions}
\usage{
global_sessions_config(
  conn = NULL,
  app_name = NULL,
  firebase_project_id = NULL,
  authorization_level = "app",
  admin_mode = FALSE,
  is_invite_required = TRUE
)
}
\arguments{
\item{conn}{the database connection}

\item{app_name}{the name of the app.}

\item{firebase_project_id}{the Firebase project ID.}

\item{authorization_level}{either "app" or "all".  Use "app" to individually authorize users to this
app.  Use "all" to give all user that have access to any of your apps access to this app.  "all" is
used by our "apps_dashboard_*".}

\item{admin_mode}{FALSE by default.  Set to TRUE to enter the polished Admin Panel without needing
to register and sign in.  This is useful during development for inviting the first users to your app.
Make sure to set `admin_mode` to FALSE before deploying your app.}

\item{is_invite_required}{TRUE by default.  Whether or not to require the user to have an
invite before registering/signing in}
}
\description{
configuration for global sessions
}
