% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_module.R
\name{profile_module}
\alias{profile_module}
\title{Profile Module Server}
\usage{
profile_module(input, output, session)
}
\arguments{
\item{input}{the Shiny server \code{input}}

\item{output}{the Shiny server \code{output}}

\item{session}{the Shiny server \code{session}}
}
\value{
\code{invisible(NULL)}
}
\description{
The server logic to accompany the  \code{\link{profile_module_ui}}.
}
