% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers_ui.R
\name{providers_ui}
\alias{providers_ui}
\title{UI for the Firebase authentication providers buttons}
\usage{
providers_ui(ns, sign_in_providers = c("google", "email"))
}
\arguments{
\item{ns}{the 'shiny' namespace function created with \code{shiny::NS()}.}

\item{sign_in_providers}{the sign in providers to enable.  Valid values are "google"
"email", "microsoft", and/or "facebook". Defaults to \code{c("google", "email")}.}
}
\value{
the html UI of the "Sign in with *" buttons.
}
\description{
Creates the html UI of the "Sign in with *" buttons.  These buttons are only
necessary if you enable social sign in via the \code{sign_in_providers} argument
passed to \code{\link{global_sessions_config}}.
}
