% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_query_string}
\alias{remove_query_string}
\title{Remove the URL query}
\usage{
remove_query_string(
  session = shiny::getDefaultReactiveDomain(),
  mode = "replace"
)
}
\arguments{
\item{session}{the Shiny \code{session}}

\item{mode}{the mode to pass to \code{shiny::updateQueryString()}.  Valid values are
\code{"replace"} or \code{"push"}.}
}
\description{
Remove the entire query string from the URL.  This function should only be called
inside the server function of your Shiny app.
}
