% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_render.R
\name{secure_render}
\alias{secure_render}
\title{Render and secure an Rmarkdown document}
\usage{
secure_render(
  rmd_file_path,
  global_sessions_config_args = list(api_key = get_api_key()),
  sign_out_button = shiny::actionLink("sign_out", "Sign Out", icon =
    shiny::icon("sign-out-alt"), class = "polished_sign_out_link")
)
}
\arguments{
\item{rmd_file_path}{the path the to .Rmd file.}

\item{global_sessions_config_args}{arguments to be passed to \code{\link{global_sessions_config}}.}

\item{sign_out_button}{action button or link with \code{inputId = "sign_out"}. Set to \code{NULL} to not include a sign out button.}
}
\value{
a Shiny app object
}
\description{
\code{secure_render()} can be used to render and secure any Rmarkdown document.
Rendering is handled by \code{rmarkdown::render} and the then the rendered document
is secured with \code{polished} authentication.
}
