% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_server.R
\name{secure_server}
\alias{secure_server}
\title{secure_server}
\usage{
secure_server(server, firebase_functions_url, app_name)
}
\arguments{
\item{server}{A Shiny server function. e.g `function(input, output, session) {}`}

\item{firebase_functions_url}{url for the firebase functions}

\item{app_name}{the name of the app}
}
\value{
session session object with new reactive session$userData$current_user which
is set to NULL if user is not signed in or a list with user data if the user is
signed in
}
\description{
secure_server
}
