% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_server.R
\name{secure_server}
\alias{secure_server}
\title{Secure your Shiny app's server}
\usage{
secure_server(server, custom_sign_in_server = NULL, custom_admin_server = NULL)
}
\arguments{
\item{server}{A Shiny server function (e.g \code{function(input, output, session) {}})}

\item{custom_sign_in_server}{Either \code{NULL}, the default, or a Shiny server containing your custom
sign in server logic.}

\item{custom_admin_server}{Either \code{NULL}, the default, or a Shiny server function containing your custom admin
server functionality.}
}
\description{
This function is used to secure your Shiny app's server function.  Make sure to pass
your Shiny app's server function as the first argument to \code{secure_server()} at
the bottom of your Shiny app's \code{server.R} file.
}
