% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_server.R
\name{secure_server}
\alias{secure_server}
\title{secure_server}
\usage{
secure_server(input, session, firebase_function_url, app_name,
  dev_user = list(email = "andy.merlino@tychobra.com", is_admin = TRUE,
  role = ""))
}
\arguments{
\item{input}{input argument from shiny server}

\item{session}{session argument from shiny server}

\item{firebase_function_url}{url for the firebase function to sign in}

\item{dev_user}{list defining the user to sign in as during development}
}
\value{
session session object with new reactive session$userData$current_user which
is set to NULL if user is not signed in or a list with user data if the user is
signed in
}
\description{
secure_server
}
