% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_static.R
\name{secure_static}
\alias{secure_static}
\title{secure_static}
\usage{
secure_static(html_file_path, global_sessions_config_args)
}
\arguments{
\item{html_file_path}{the path the to html file.  See the details for more info.}

\item{global_sessions_config_args}{arguments to be passed to \code{\link{global_sessions_config}}.}
}
\value{
a Shiny app object
}
\description{
Secure a static html page.  \code{secure_static()} can be used to secure any html
page using 'polished'.  It is often used to add 'polished' to ".Rmd" htmloutput
and flexdashboards.
}
\details{
To secure a static html page, place the html page in a folder named "www"
and call \code{secure_static()} from a file named "app.R".  The file structure should
look like:
\itemize{
\item app.R
\item www/
\itemize{
\item index.html
}
}

See an example here: \url{https://github.com/Tychobra/polished_example_apps/tree/master/05_flex_dashboard}
}
