% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_ui.R
\name{secure_ui}
\alias{secure_ui}
\title{Secure the Shiny Application UI}
\usage{
secure_ui(ui, firebase_config, app_name, sign_in_page_ui = NULL,
  custom_admin_ui = NULL)
}
\arguments{
\item{ui}{UI of the application.}

\item{firebase_config}{A list containing your Firebase project configuration.  This will
be passed to \code{\link{firebase_init()}}.}

\item{app_name}{The name of the app.}

\item{sign_in_page_ui}{Either `NULL`, the default, or the HTML, CSS, and JavaScript
to use for the UI of the Sign In page.}

\item{custom_admin_ui}{Either `NULL`, the default, or a list of 2 elements containing custom
ui to add addtional `shinydashboard` tabs to the Polished admin panel.}
}
\value{
Secured Shiny app ui
}
\description{
Secure the Shiny Application UI
}
