% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_ui.R
\name{secure_ui}
\alias{secure_ui}
\title{secure_ui}
\usage{
secure_ui(
  ui,
  sign_in_page_ui = NULL,
  custom_admin_ui = NULL,
  custom_admin_button_ui = admin_button_ui("polished"),
  admin_ui_options = default_admin_ui_options()
)
}
\arguments{
\item{ui}{UI of the application.}

\item{sign_in_page_ui}{Either `NULL`, the default, or the HTML, CSS, and JavaScript
to use for the UI of the Sign In page.}

\item{custom_admin_ui}{Either `NULL`, the default, or a list of 2 elements containing custom
ui to add addtional `shinydashboard` tabs to the Polished admin panel.}

\item{custom_admin_button_ui}{Either `admin_button_ui("polished")`, the default, ot your custom
ui to take admins from the custom Shiny app to the Admin panel.}

\item{admin_ui_options}{list of custom UI options.  Passed as argument to `admin_module_ui()`.}
}
\value{
Secured Shiny app ui
}
\description{
This function is used to secure your 'shiny' app's UI.  Make sure to pass
your 'shiny' app's UI as the first argument to \code{secure_ui()} at
the bottom of your 'shiny' app's "ui.R" file.
}
