% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_ui.R
\name{secure_ui}
\alias{secure_ui}
\title{Secure the Shiny Application UI}
\usage{
secure_ui(ui, firebase_config, sign_in_page_ui = NULL,
  custom_admin_ui = NULL,
  custom_admin_button_ui = admin_button_ui("polished"))
}
\arguments{
\item{ui}{UI of the application.}

\item{firebase_config}{Firebase configuration.}

\item{sign_in_page_ui}{Either `NULL`, the default, or the HTML, CSS, and JavaScript
to use for the UI of the Sign In page.}

\item{custom_admin_ui}{Either `NULL`, the default, or a list of 2 elements containing custom
ui to add addtional `shinydashboard` tabs to the Polished admin panel.}

\item{custom_admin_button_ui}{Either `admin_button_ui("polished")`, the default, ot your custom
ui to take admins from the custom Shiny app to the Admin panel.}
}
\value{
Secured Shiny app ui
}
\description{
Secure the Shiny Application UI
}
