% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_password_reset_email_module.R
\name{send_password_reset_email_module}
\alias{send_password_reset_email_module}
\title{the server logic for a Shiny module to send a password reset email}
\usage{
send_password_reset_email_module(input, output, session, email)
}
\arguments{
\item{input}{the Shiny server \code{input}}

\item{output}{the Shiny server \code{output}}

\item{session}{the Shiny server \code{session}}

\item{email}{A reactive value returning the email address to send the password
reset email to.}
}
\value{
\code{invisible(NULL)}
}
\description{
This function sends a request to the \url{https://polished.tech} API to reset a user's
password.
}
