% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_password_reset_email_module.R
\name{send_password_reset_email_module_ui}
\alias{send_password_reset_email_module_ui}
\title{the UI for a Shiny module to send a password reset email}
\usage{
send_password_reset_email_module_ui(id, link_text = "Forgot your password?")
}
\arguments{
\item{id}{the Shiny module \code{id}}

\item{link_text}{text to use for the password reset link.}
}
\description{
the UI for a Shiny module to send a password reset email
}
