% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sessions.R
\docType{data}
\name{Sessions}
\alias{Sessions}
\title{Sessions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Sessions
}
\description{
R6 class to track the polished sessions
}
\keyword{datasets}
