% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sessions.R
\name{Sessions}
\alias{Sessions}
\title{Sessions}
\description{
R6 class to track the polished sessions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-config}{\code{Sessions$config()}}
\item \href{#method-sign_in}{\code{Sessions$sign_in()}}
\item \href{#method-get_invite_by_email}{\code{Sessions$get_invite_by_email()}}
\item \href{#method-get_invite_by_uid}{\code{Sessions$get_invite_by_uid()}}
\item \href{#method-get_roles}{\code{Sessions$get_roles()}}
\item \href{#method-find}{\code{Sessions$find()}}
\item \href{#method-remove}{\code{Sessions$remove()}}
\item \href{#method-list}{\code{Sessions$list()}}
\item \href{#method-refresh_email_verification}{\code{Sessions$refresh_email_verification()}}
\item \href{#method-log_session}{\code{Sessions$log_session()}}
\item \href{#method-set_signed_in_as}{\code{Sessions$set_signed_in_as()}}
\item \href{#method-clear_signed_in_as}{\code{Sessions$clear_signed_in_as()}}
\item \href{#method-clone}{\code{Sessions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-config"></a>}}
\subsection{Method \code{config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$config(
  app_name,
  firebase_functions_url = NULL,
  conn = NULL,
  authorization_level = "app"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sign_in"></a>}}
\subsection{Method \code{sign_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$sign_in(firebase_token, token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_invite_by_email"></a>}}
\subsection{Method \code{get_invite_by_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$get_invite_by_email(email)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_invite_by_uid"></a>}}
\subsection{Method \code{get_invite_by_uid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$get_invite_by_uid(user_uid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_roles"></a>}}
\subsection{Method \code{get_roles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$get_roles(user_uid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find"></a>}}
\subsection{Method \code{find()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$find(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$remove(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh_email_verification"></a>}}
\subsection{Method \code{refresh_email_verification()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$refresh_email_verification(token, firebase_uid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_session"></a>}}
\subsection{Method \code{log_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$log_session(token, user_uid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_signed_in_as"></a>}}
\subsection{Method \code{set_signed_in_as()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$set_signed_in_as(token, signed_in_as)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_signed_in_as"></a>}}
\subsection{Method \code{clear_signed_in_as()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$clear_signed_in_as(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sessions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
