% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config_env.R
\name{set_config_env}
\alias{set_config_env}
\title{set_config_env}
\usage{
set_config_env(override = NULL)
}
\arguments{
\item{override}{Set the environment to "default" or "production" manully. \strong{CAUTION:}
Be sure you know the difference between "default" & "production" configuration environments.
Using the "production" environment will affect the database of the deployed application.}
}
\description{
Determines if the app is deployed on shinyapps.io, where the environment variable USER is 'shiny', 
or running locally and adjusts the config environment accordingly
}
