% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config_env.R
\name{set_config_env}
\alias{set_config_env}
\title{Automatically set the config environment}
\usage{
set_config_env(override = NULL)
}
\arguments{
\item{override}{Set the environment to \code{"default"} or \code{"production"} manually. \strong{CAUTION:}
Be sure you know the difference between \code{"default"} & \code{"production"} configuration environments.
Using the \code{"production"} environment will affect the database of the deployed application.}
}
\value{
\code{invisible(NULL)}
}
\description{
Determines if the app is deployed to a server or running locally, and adjusts
the config environment to \code{"production"} or \code{"default"}, respectively.  This function
is almost always called in the \code{global.R} file of a Shiny app immediately before
the configuration in the \code{config.yml} is read in.
}
