% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config_env.R
\name{set_config_env}
\alias{set_config_env}
\title{set_config_env}
\usage{
set_config_env(session, override = NULL)
}
\arguments{
\item{override}{Set the environment to "default" or "production" instead of 
the value found through session$clientData$hostname}
}
\description{
Determines if the app is deployed or running locally and adjusts the
config environment accordingly
}
