% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_module.R
\name{sign_in_module}
\alias{sign_in_module}
\title{Server logic for the sign in and register pages}
\usage{
sign_in_module(input, output, session)
}
\arguments{
\item{input}{the Shiny input}

\item{output}{the Shiny output}

\item{session}{the Shiny session}
}
\description{
This server logic accompanies the \code{\link{sign_in_module_ui}}.
}
