% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_module.R
\name{sign_in_module_ui}
\alias{sign_in_module_ui}
\title{sign_in_module_ui}
\usage{
sign_in_module_ui(id, register_link = "First time user? Register here!")
}
\arguments{
\item{id}{the Shiny module id}

\item{register_link}{The text that will be displayed in the link to go to the
user registration page.  The default is "First time user? Register here!".
Set to \code{NULL} if you don't want to use the registration page.}
}
\description{
UI for the sign in and register pages when a user invite is required to register and
sign in.  See \code{\link{sing_in_no_invite_module}} if you do not require your
users to sign in and register to access your 'shiny' app.
}
