% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_no_invite_module.R
\name{sign_in_no_invite_module_ui}
\alias{sign_in_no_invite_module_ui}
\title{sign_in_no_invite_module_ui}
\usage{
sign_in_no_invite_module_ui(id)
}
\arguments{
\item{id}{the Shiny module id}
}
\description{
UI for the sign in and register panels.  User is not required to have an invite
to sign in and register.
}
