% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in_ui_default.R
\name{sign_in_ui_default}
\alias{sign_in_ui_default}
\title{sign_in_ui_default}
\usage{
sign_in_ui_default(
  sign_in_module = sign_in_module_ui("sign_in"),
  color = "#AAA",
  company_name = "Your Brand Here",
  logo_top = div(class = "text-center", style = "color: #FFF;", h1("Your", style =
    "margin-bottom: 0; margin-top: 30px;"), h1("Brand", style =
    "margin-bottom: 0; margin-top: 10px;"), h1("Here", style =
    "margin-bottom: 15px; margin-top: 10px;")),
  logo_bottom = tags$img(src = "polish/images/placeholder_company_logo.jpg", alt =
    "Placeholder Logo", style = "width: 200px; margin-bottom: 15px; padding-top: 15px;"),
  icon_href = "polish/images/polished_icon.png"
)
}
\arguments{
\item{sign_in_module}{UI module for the sign-in page}

\item{color}{hex color for the background and button}

\item{company_name}{your company name}

\item{logo_top}{html for logo to go above the sign in panel.}

\item{logo_bottom}{html for the logo below the sign in panel.}

\item{icon_href}{the url/path to the browser tab icon}
}
\value{
the UI for the sign in page
}
\description{
sign_in_ui_default
}
