% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_out.R
\name{sign_out}
\alias{sign_out}
\title{sign_out}
\usage{
sign_out(conn, hashed_cookie, session_uid, schema = "polished")
}
\arguments{
\item{conn}{the database connection}

\item{hashed_cookie}{the user's hashed browser cookie}

\item{session_uid}{the user's session uid}

\item{schema}{the database schema}
}
\description{
sign_out
}
