% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_out_from_shiny.R
\name{sign_out_from_shiny}
\alias{sign_out_from_shiny}
\title{sign_out_from_shiny}
\usage{
sign_out_from_shiny(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{the Shiny session}
}
\description{
sign_out_from_shiny
}
