% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_out_from_shiny.R
\name{sign_out_from_shiny}
\alias{sign_out_from_shiny}
\title{sign_out_from_shiny}
\usage{
sign_out_from_shiny(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{the Shiny session}
}
\description{
Call this function to sign a user out of your 'shiny' app.  This function should
be called inside the server function of your 'shiny' app.  See
\url{https://github.com/Tychobra/polished/blob/master/inst/examples/polished_example_01/server.R}
For an example of this function being called after the user clicks a "Sign Out"
button.
}
