% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polished_api_01_apps.R
\name{update_app}
\alias{update_app}
\title{Polished API - Update an App}
\usage{
update_app(
  app_uid = NULL,
  app_name = NULL,
  app_url = NULL,
  api_key = getOption("polished")$api_key
)
}
\arguments{
\item{app_uid}{the app uid to update.}

\item{app_name}{an optional app name.}

\item{app_url}{an optional app url.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Update an App
}
\details{
supply either the app uid or app name to get data about a specific app.
}
\seealso{
get_apps add_app delete_app
}
