% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_03_app_users.R
\name{update_app_user}
\alias{update_app_user}
\title{Polished API - Update an App}
\usage{
update_app_user(
  app_uid,
  user_uid,
  is_admin,
  api_key = getOption("polished")$api_key
)
}
\arguments{
\item{app_uid}{the app uid to update.}

\item{user_uid}{an optional app name.}

\item{is_admin}{boolean - whether or not the user is an admin.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key()}}
so that you do not need to supply this argument with each function call.}
}
\description{
Polished API - Update an App
}
\seealso{
get_app_users add_app_user delete_app_user
}
