% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_user.R
\name{update_app_user}
\alias{update_app_user}
\title{update app user}
\usage{
update_app_user(
  conn,
  user_uid,
  app_uid,
  is_admin,
  modified_by,
  schema = "polished"
)
}
\arguments{
\item{conn}{the database connection}

\item{user_uid}{the uid for the user to update}

\item{app_uid}{the app id}

\item{is_admin}{boolean whether or not the user is and admin}

\item{modified_by}{the uid of the user making the edit}

\item{schema}{the database schema}
}
\description{
Update a user authorized to access and app.
}
