% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_06_sessions.R
\name{update_session}
\alias{update_session}
\title{Polished API - Update a session}
\usage{
update_session(session_uid, dat, api_key = get_api_key())
}
\arguments{
\item{session_uid}{the role uid of the role to be deleted.}

\item{api_key}{your Polished API key.  Set your polished api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}

\item{user_uid}{the user uid that the role should be removed from.}
}
\description{
Polished API - Update a session
}
\seealso{
\code{\link[=get_user_roles]{get_user_roles()}} \code{\link[=add_user_role]{add_user_role()}}
}
