% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\docType{data}
\name{User}
\alias{User}
\title{User}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
User
}
\description{
R6 class to track the polish user.  An instance of this class named `polish__user` should
be created in "global.R" of the Shiny app.
}
\keyword{datasets}
