% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_email_module.R
\name{verify_email_module_ui}
\alias{verify_email_module_ui}
\title{verify email page ui}
\usage{
verify_email_module_ui(id, firebase_config)
}
\arguments{
\item{id}{the Shiny module id}

\item{firebase_config}{Firebase configuration}
}
\description{
verify email page ui
}
